fix: move lib to api
